+++ examples/c99/track_beambeam.c
@@ -70,7 +70,8 @@ int main( int argc, char* argv[] )
     }
 
     printf( "Use: NUM_PARTICLES = %10lu\r\n"
-            "     NUM_TURNS     = %10lu\r\n\r\n", NUM_PARTICLES, NUM_TURNS );
+            "     NUM_TURNS     = %10lu\r\n\r\n",
+            ( unsigned int )NUM_PARTICLES, ( unsigned int )NUM_TURNS );
 
     /* ********************************************************************** */
     /* ****   Building Particles Data from Input Example Particle Data   **** */
